Revert "remove CacheWarmer because it doesn't make much sense to recompile all"
it actually does make sense because assetic also warms up its cache and
fails to do so beacuse of missing asset files referenced in the twig
templates and would not dump them with assetic:dump later on due to a
non-valid cache. it is therefore required to either do the following order

cache:clear --no-warmup
compassomator:compile
cache:warmup

which would fail on composer updates/installs or just put in the cache
warmer before the twig/assetic cache warmer to have the least surprises.

This reverts commit e9210291ce9dd80074790d95c73f5974788dd59b.